+++ src/pine/ast/table.clj
@@ -1,7 +1,6 @@
 (ns pine.ast.table
   (:require
-   [pine.db.main :as db]
-   [pine.hints :as hints]))
+   [pine.db.main :as db]))
 
 (defn- join-helper [references t1 t2 a1 a2 c direction]
   "Find the references between the tables, get the columns for the first
@@ -29,23 +28,20 @@
                 (join-helper references t2 t1 a2 a1 c :reverse))]
     result))
 
-(defn- update-joins [state {:keys [join-column parent]}]
+(defn- update-joins [state current]
   "Use the tables in the state to create a join between the last 2 tables. The
   reason to get the tables from the state is that they have been assigned an
   alias. We only use the join column from the current value being processed."
-  (let [{:keys [tables joins]} state]
+  (let [{:keys [join-column parent]} current
+        from-alias                   (state :context)
+        {:keys [tables join-map]}       state]
     (cond
-      (< (count tables) 2) state
-      :else
-      (let [[x y] (take-last 2 tables)
-            join-result (join state x y join-column parent)]
-        (assoc-in state [:joins (x :alias) (y :alias)] join-result)))))
-
-(defn- update-hints [state value]
-  (let [hints (if (= (state :pending-count) 1)
-                (hints/generate state value)
-                {:table []})]
-    (assoc-in state [:hints] hints)))
+      (nil? from-alias) state
+      :else (let [x (-> state :aliases (get from-alias))
+                  join-result (join state x current join-column parent)]
+              (-> state
+                  (assoc-in [:join-map (x :alias) (current :alias)] join-result)
+                  (update :joins conj [(x :alias) (current :alias) join-result]))))))
 
 (defn make-alias [s]
   (let [words (if (not-empty s) (clojure.string/split s #"_") ["x"])
@@ -55,16 +51,20 @@
 ;; todo: spec for the :value for a :table
 (defn handle [state value]
   (let [{:keys [table alias schema parent]} value
-        a (or alias (str (make-alias table) "_" (state :table-count)))]
+        a (or alias (str (make-alias table) "_" (state :table-count)))
+        current {:schema schema :table table :alias a :parent parent}]
     (-> state
-        (update :tables conj {:schema schema :table table :alias a :parent parent})
-        (update :aliases assoc a {:table table :schema schema})
-        (update-joins value)
-        (update-hints value)
+        ;; pre
+        (assoc  :context (state :current))
+        (assoc  :current a)
+        ;;
+        (update :tables conj current)
+        (update :aliases assoc a current)
+        (update-joins current)
+        ;; post
         ;;
         ;; TODO: These are metadata - mayebe I should move them to a different
         ;; ns e.g. if the operation is table, then I update the following there
         ;;
-        (assoc  :context a)
         (update :table-count inc))))