+++ src/pine/db/fixtures.clj
@@ -17,15 +17,16 @@
 ;;              |       |                       |
 ;;              |       |       id (PK)         |
 ;;              +-------|      company_id (FK)  |<---|
-;;                      |      reports_to (FK)  |----|
-;;                      +-----------------------+
-;;                          ^
-;;                          |       +------------------------+
-;;                          |       |       z.document       |
-;;                          |       |                        |
-;;                          |       |      id (PK)           |
-;;                          |-------|     employee_id (FK)   |
-;;                          +-------|     created_by  (FK)   |
+;;              |       |      reports_to (FK)  |----|
+;;              |       +-----------------------+
+;;              |           ^
+;;              |           |       +------------------------+
+;;              |           |       |       z.document       |
+;;              |           |       |                        |
+;;              |           |       |      id (PK)           |
+;;              |           |-------|     employee_id (FK)   |
+;;              |           +-------|     created_by  (FK)   |
+;;              +-------------------|     company_id (FK)    |
 ;;                                  +------------------------+
 
 (def references [["y"  "employee"      "company_id"    "x"  "company"  "id"]
@@ -33,5 +34,7 @@
                  ["z"  "document"      "created_by"    "y"  "employee" "id"]
 
                  ;; self join
-                 ["y"  "employee"      "reports_to"   "y"  "employee" "id"]])
+                 ["y"  "employee"      "reports_to"    "y"  "employee" "id"]
+
+                 ["z"  "document"      "company_id"    "x"  "company" "id"]])